Competition Code Version 1.0

+++ src/com/shwinlib/ShwinPID.java
@@ -3,14 +3,14 @@ package com.shwinlib;
 public class ShwinPID {
   long deltaT;
   long prevT;
-  float kP;
-  float kI;
-  float kD;
-  float integLimit;
-  float integData;
-  float prevError = 0, prevTime = 0;
+  double kP;
+  double kI;
+  double kD;
+  double integLimit;
+  double integData;
+  double prevError = 0, prevTime = 0;
   
-  public ShwinPID(float kP, float kI, float kD, float integLimit) {
+  public ShwinPID(double kP, double kI, double kD, double integLimit) {
     this.kP = kP;
     this.kI = kI;
     this.kD = kD;
@@ -20,14 +20,14 @@ public class ShwinPID {
  
 
 
-  public float doPID(float error) {
+  public double doPID(double error) {
     deltaT = (long) (System.currentTimeMillis() - prevTime);
     prevTime = System.currentTimeMillis();
-    float P = error * kP;
-    float I = 0.0f; //you dont need this right now
-    float D = (error-prevError)/deltaT * kD;
+    double P = error * kP;
+    double I = 0.0f; //you dont need this right now
+    double D = (error-prevError)/deltaT * kD;
     prevError = error;
-    return (float)(P + I + D); 
+    return (P + I + D); 
   }
